Moved G-Collector to the libft module + All std libs are included only via ./includes.h file
