+++ libft/memory.h
@@ -1,7 +1,7 @@
 #ifndef MEMORY_H
 # define MEMORY_H
 
-# include "./types.h"
+# include "../types.h"
 
 void	*ft_malloc(size_t	size, size_t unit);
 void	ft_gcadd_back(void *addr);
